Add files via upload
This simple function allows you to mark the optic nerve head boundaries on OCT image of the retina. Can be modified to mark any number of points. 
The additional fit_ellipse function takes in the marked points, and calculates the ONH parameters such as the diameter, area, etc. This additional function and a sample image is also included in this repository.